Python/requirements.txt
# Python PIP (package manager) required dependencies
# Run `pip install -r requirements.txt` to install the required packages.
PyHamcrest==1.9.0
six==1.11.0
# six is dependency of PyHamcrest
mock==2.0.0
